refactor: 클린아키텍처 구조 반영(#7)
