+++ app/adapter/input/graphql/exception_handler.py
