+++ app/application/use_cases/bus_stop.py
@@ -1,18 +1,52 @@
 from abc import ABC, abstractmethod
-from fastapi import Depends
-from typing import List
+from typing import List, Optional
+from pydantic import BaseModel
 
-from app.dtos.request.bus_stop import GetBusStopRequestDTO
-from app.dtos.request.bus_stop import SearchBusStopsRequestDTO
-from app.dtos.response.bus_stop import GetBusStopResponseDTO
-from app.dtos.response.bus_stop import SearchBusStopsResponseDTO
+from app.application.repositories.bus_stop import BusStopRepository
+from app.application.repositories.bus_route import BusRouteRepository
+
+
+class GetBusStopRequestDTO(BaseModel):
+    id: int
+    requested_field: List[str]
+
+
+class SearchBusStopsRequestDTO(BaseModel):
+    name: Optional[str]
+    lat: Optional[float]
+    lng: Optional[float]
+    top_k: int
+    requested_fields: List[str]
+
+
+class GetBusStopResponseDTO(BaseModel):
+    id: Optional[int]
+    name: Optional[str]
+    lat: Optional[float]
+    lng: Optional[float]
+    arrivals: Optional[List[dict]]
+    stop_bus_routes: Optional[List[dict]]
+
+
+class SearchBusStopsResponseDTO(BaseModel):
+    bus_stops: List[GetBusStopResponseDTO]
 
 
 class BusStopUseCase(ABC):
     @abstractmethod
-    def find_by_id(dto: GetBusStopRequestDTO) -> GetBusStopResponseDTO:
+    def __init__(
+        self,
+        bus_stop_repository: BusStopRepository,
+        bus_route_repository: BusRouteRepository,
+    ):
         pass
 
     @abstractmethod
-    def search_bus_stops(dto: SearchBusStopsRequestDTO) -> SearchBusStopsResponseDTO:
+    def get_bus_stop(self, dto: GetBusStopRequestDTO) -> GetBusStopResponseDTO:
+        pass
+
+    @abstractmethod
+    def search_bus_stops(
+        self, dto: SearchBusStopsRequestDTO
+    ) -> SearchBusStopsResponseDTO:
         pass
