+++ app/main.py
@@ -3,8 +3,8 @@ from strawberry.fastapi import GraphQLRouter
 from contextlib import asynccontextmanager
 import asyncio
 
-from app.interfaces.input.v1.graphql.schema import schema
-from app.interfaces.input.v1.graphql.dependencies import get_context
+from app.adapter.input.graphql.schema import schema
+from app.adapter.input.graphql.context import get_context
 
 
 @asynccontextmanager
@@ -23,5 +23,5 @@ async def lifespan(app: FastAPI):
 app = FastAPI(docs_url="/docs", openapi_url="/open-api-docs", lifespan=lifespan)
 
 # Strawberry GraphQL Router 연결
-graphql_app = GraphQLRouter(schema, context_getter=get_context)
-app.include_router(graphql_app, prefix="/api/v1/graphql")
+graphql_app = GraphQLRouter(schema=schema, context_getter=get_context)
+app.include_router(graphql_app, prefix="/graphql")
